Separate `output!` function from `postprocess!`
This provides more granularity to the client over compute vs data
transfer.  There is still some compute in `output!`, but it is an FFT
that outputs directly to a user supplied output buffer.  If this proves
problematic/suboptimal the FFT output buffer could live in the
`ZDTWorkspace`.  This would make the `output!` function purely data
transfer (i.e. no compute), but it would increase the memory footprint
in terms of both memory space and memory bandwidth utilization.