Initialize ZDTWorkspace V field in constructor
Any V locations that fall between vlow and vhigh are described as "don't
care" values, but actually they are "don't care so long as they are not
NaN" values, so we need to initialize them to non-NaN values.  The
easiest way to do that is to initialize all of V to zero.  This happens
only once when the workspace is constructed so it's not too critical to
limit the region that gets initialized, but there is room for a little
optimization if needed.