Add CuArray-aware `output!` method in extension
C2R transforms have a CUDA.jl hack to work around a CUFFT "known issue"
that "“cuFFT will always overwrite the input for out-of-place C2R
transform”.  The hack duplicates the input, does the FFT, then frees the
copy.  This can be problematic when getting close(-ish) to a memory full
condition.  This commit bypasses the copy workaround so that the plan
can run without copying (or, more importantly, allocating) any
extraneous data.