Invert the sign convention of drift rate
The normalized drift rates used in this package now refer to the rate at
which a signal is drifting rather than the drift correction rate that
should be applied to the siganal to align it in frequency.

In prior versions, the normalized drift rates used in this package
referred to the amount of shifting over time required to be applied to a
drifting signal to align that signal in frequenecy.

This is obviosuly a breaking change, but the new sign convention is more
intuitive and compatible with other existing Doppler drift rate tools.