Add fdrsynchronize(::Type) function
The `fdrsynchronize` function should be called whenever synchronization
might be needed to ensure that any work that has been enqueued in one
Task completes before passing a resource (such as an `Array` or
`CuArray`) to another Task.  When using `Arrays` (which live in CPU
memory and do not support enqueueing work), this function does nothing.
When using `CuArrays` (which live in GPU memory and do support
enqueueing work), this function calls `CUDA.synchronize()`.